Added a level with new mechanics
I added a new level with a new dive mechanic, if the player attempts to dive in a location that's not allowed (above or below a rock) a screen shake effect is played from Emily's Camera Pivot class. Beyond this, I also added a basic drum sound effect to imply that it won't work. Once my branch and Emily's branch is pulled into main, my LevelController class should begin to work properly.